Add execute bit to directories before bundling
