+++ src/lib/package-generator.ts
@@ -1,9 +1,12 @@
-import { AnyPackage, TypesDataFile, TypingsData, NotNeededPackage, fullPackageName, notNeededReadme, settings, getOutputPath } from "./common";
-import { Log, Logger, quietLogger } from "./logging";
-import { hasOwnProperty, readFile, readJson, writeFile } from "./util";
-import Versions, { VersionInfo } from "./versions";
 import * as fsp from "fs-promise";
 import * as path from "path";
+
+import { readFile, readJson, writeFile } from "../util/io";
+import { Log, Logger, quietLogger } from "../util/logging";
+import { hasOwnProperty } from "../util/util";
+
+import { AnyPackage, TypesDataFile, TypingsData, NotNeededPackage, fullPackageName, notNeededReadme, settings, getOutputPath } from "./common";
+import Versions, { VersionInfo } from "./versions";
 
 /** Generates the package to disk */
 export default function generateAnyPackage(pkg: AnyPackage, availableTypes: TypesDataFile, versions: Versions): Promise<Log> {
