+++ src/test-get-secrets.ts
@@ -1,7 +1,7 @@
 // Run `node ./bin/test-get-secrets.js` to test that we can fetch secrets from Azure Keyvault
 
 import { allSecrets, getSecret, Secret } from "./lib/secrets";
-import { done } from "./lib/util";
+import { done } from "./util/util";
 
 if (!module.parent) {
 	done(main());
